Add files via upload
This project is using python to conduct analysis of the logistic warehouse system in Walmart for the incoming Christmas. The project is broken into two parts: 1. Selecting products for each warehouse according to the value-weight ratio before the promotion starts; 2. Simulating cross-warehouse-transshipment solution for possible product shortage problem for a random warehouse after the promotion starts. I mainly used NumPy package in Python for this project.